+++ inbox/lib/constants.dart
@@ -1,4 +1,4 @@
-const String TAG = "MoeInbox_";
+const String TAG = "Inbox_";
 const String CHANNEL_NAME = "com.moengage/inbox";
 
 const String METHOD_NAME_UN_CLICKED_COUNT = "unClickedCount";