+++ inbox/lib/moe_ios_inbox.dart
@@ -1,13 +1,9 @@
-import 'dart:convert';
 import 'dart:core';
 
 import 'package:flutter/services.dart';
 import 'package:moengage_inbox/inbox_data.dart';
-import 'package:moengage_inbox/inbox_message.dart';
 import 'package:moengage_inbox/payload_transformer.dart';
 import 'package:moengage_inbox/constants.dart';
-
-import 'payload_transformer.dart';
 
 class MoEiOSInbox {
   MethodChannel _channel;